xen/arm: fix affected memory range by dcache clean functions
clean_dcache_va_range and clean_and_invalidate_dcache_va_range don't
calculate the range correctly when "end" is not cacheline aligned. As a
result, the last cacheline is not skipped. Fix the issue by aligning the
start address to the cacheline size.

In addition, make the code simpler and faster in
invalidate_dcache_va_range, by removing the module operation and using
bitmasks instead.

Signed-off-by: Stefano Stabellini <sstabellini@kernel.org>
Reported-by: edgar.iglesias@xilinx.com
CC: edgar.iglesias@xilinx.com