Tools/libxc: Add viommu operations in libxc
In previous patch, we introduce a common vIOMMU layer. In our design,
we create/destroy vIOMMU through DMOP interface instead of creating it
according to a config flag of domain. It makes it is possible
to create vIOMMU in device model or in tool stack.

The following toolstack code is to add XEN_DMOP_viommu_XXX syscalls:
 - query capabilities of vIOMMU emulated by Xen
 - create vIOMMU in Xen hypervisor with base address, capability
 - destroy vIOMMU specified by viommu_id

Signed-off-by: Chao Gao <chao.gao@intel.com>
Signed-off-by: Lan Tianyu <tianyu.lan@intel.com>