tools: reserve guest memory for ACPI from device model
Some virtual devices (e.g. NVDIMM) require complex ACPI tables and
definition blocks (in AML), which a device model (e.g. QEMU) has
already been able to construct. Instead of introducing the similar
implementation to Xen, we would like to reuse the device model to
provide those ACPI stuffs.

This commit reserves an area in the guest memory for the device model
to pass its ACPI tables and definition blocks to guest, which will be
loaded by hvmloader. The base guest physical address and the size of
the reserved area are passed to the device model via XenStore keys
hvmloader/dm-acpi/{address, length}.

Signed-off-by: Haozhong Zhang <haozhong.zhang@intel.com>