tools/libxl: add support for emulated NVMe drives
Upstream QEMU supports emulation of NVM Express a.k.a. NVMe drives.

This patch adds a new vdev type into libxl to allow such drives to be
presented to HVM guests. Because the purpose of the new vdev is purely
to configure emulation, the syntax only supports specification of
whole disks. Also there is no need to introduce a new concrete VBD
encoding for NVMe drives.

NOTE: QEMU's emulation only supports a single NVMe namespace, so the
      vdev syntax does not include specification of a namespace.
      Also, current versions of SeaBIOS do not support booting from
      NVMe devices, so the vdev should only be used for secondary
      drives.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>