libxl/xl: add support for Xen 9pfs
Add functions to libxl to setup a Xen 9pfs frontend/backend connection.
Add support to xl to parse a xen_9pfs option in the VM config file, in
the following format:

xen_9pfs=["tag=share_dir,security_model=none,path=/path/share_dir"]

where tag identifies the 9p share and it is required to mount it on the
guest side, path is the path of the filesystem to share and the only
security_model supported is "none" which means that files are stored
using the same credentials as they are created on the guest (no user
ownership squash or remap).

Signed-off-by: Stefano Stabellini <stefano@aporeto.com>
CC: wei.liu2@citrix.com
CC: ian.jackson@eu.citrix.com