libxc/migration: correct hvm record ordering specification
The libxc migration stream specification document asserts that, within
an hvm migration stream, "HVM_PARAMS must precede HVM_CONTEXT, as
certain parameters can affect the validity of architectural state in the
context."  This sounds reasonable, but the in-tree implementation of hvm
domain save actually writes these records in the _reverse_ order, with
HVM_CONTEXT first and HVM_PARAMS next.  This has been the case for the
entire history of that implementation, seemingly to no ill effect, so
update the spec to reflect this.

Signed-off-by: Joshua Otto <jtotto@uwaterloo.ca>