Enable compiling with gcc tracing
Make minimal adjustments in order to enable the compilation of the
xen source-code with gcc-6's -fsanitize-coverage=trace-pc option.

Due to a bug in Xen's build-system the flag for the compiler has
to be handed in via the command line, i.e. for compiling one would
use:

	make CC=<path to gcc-6>

This is an experimental patch as in a final version you would not
want all files to be compiled with this option by default.

Signed-off-by: Felix Schmoll <eggi.innovations@gmail.com>